Move frontend modules into docs
