+++ docs/index.html
@@ -84,6 +84,6 @@
     </div>
   </footer>
 
-  <script type="module" src="./app.js"></script>
+  <script type="module" src="./src/app.js"></script>
 </body>
 </html>
